+++ demo/app/services/app-settings-service/app-settings-service.ts
@@ -1,20 +1,44 @@
+import { Injectable } from '@angular/core';
+
+interface AppSettings {
+  show3dtiles: boolean;
+  numberOfEntities: number;
+  entitiesUpdateRate: number;
+  showTracksLayer: boolean;
+  showMapLayer: boolean;
+  showVelocityVectors: boolean;
+  showEllipses: boolean;
+  keyboardControl: boolean;
+  tracksDataType: TracksType;
+  
+}
+
+export enum TracksType {
+	REAL_DATA,
+	SIM_DATA,
+	MODELS_3D,
+}
+
+@Injectable()
 export class AppSettingsService {
-  get settings(): { numberOfEntities: number; entitiesUpdateRate: number; showTracksLayer: boolean; showMapLayer: boolean; realTracksData: boolean; showVelocityVectors: boolean; showEllipses: boolean } {
+  get settings(): AppSettings  {
     return this._settings;
   }
 
-  set settings(value: { numberOfEntities: number; entitiesUpdateRate: number; showTracksLayer: boolean; showMapLayer: boolean; realTracksData: boolean; showVelocityVectors: boolean; showEllipses: boolean }) {
+  set settings(value: AppSettings) {
     this._settings = value;
   }
 
-  private _settings = {
+  private _settings: AppSettings = {
     numberOfEntities: 0,
     entitiesUpdateRate: 0,
     showTracksLayer: true,
     showMapLayer: true,
-    realTracksData: false,
     showVelocityVectors: false,
-    showEllipses: false
+    showEllipses: false,
+    show3dtiles: false,
+    keyboardControl: false,
+		tracksDataType: TracksType.MODELS_3D,
   };
 
   setSettings(settings) {
@@ -45,12 +69,12 @@ export class AppSettingsService {
     this._settings.showMapLayer = value;
   }
 
-  get realTracksData(): boolean {
-    return this._settings.realTracksData;
+  get tracksDataType(): TracksType {
+    return this._settings.tracksDataType;
   }
 
-  set realTracksData(value: boolean) {
-    this._settings.realTracksData = value;
+  set tracksDataType(value: TracksType) {
+    this._settings.tracksDataType = value;
   }
 
   get showTracksLayer(): boolean {
@@ -75,5 +99,21 @@ export class AppSettingsService {
 
   set numOfEntities(value: number) {
     this._settings.numberOfEntities = value;
+  }
+
+  get show3dtiles() {
+    return this._settings.show3dtiles;
+  }
+
+  set show3dtiles(value: boolean) {
+    this._settings.show3dtiles = value;
+  }
+
+  get keyboardControl() {
+    return this._settings.keyboardControl;
+  }
+
+  set keyboardControl(value: boolean) {
+    this._settings.keyboardControl = value;
   }
 }