+++ src/angular-cesium.module.ts
@@ -38,12 +38,20 @@ import { AcStaticCircleDescComponent } from './components/static-dynamic/ac-stat
 import { AcDynamicCircleDescComponent } from './components/static-dynamic/ac-dynamic-circle-desc/ac-dynamic-circle-desc.component';
 import { AcStaticPolylineDescComponent } from './components/static-dynamic/ac-static-polyline-desc/ac-static-polyline-desc.component';
 import { AcModelDescComponent } from './components/ac-model-desc/ac-model-desc.component';
+import { AcTileset3dComponent } from './components/ac-3d-tileset/ac-tileset-3d.component';
+import { AcBoxDescComponent } from './components/ac-box-desc/ac-box-desc.component';
+import { AcCylinderDescComponent } from './components/ac-cylinder-desc/ac-cylinder-desc.component';
+import { AcCorridorDescComponent } from './components/ac-corridor-desc/ac-corridor-desc.component';
+import { AcEllipsoidDescComponent } from './components/ac-ellipsoid-desc/ac-ellipsoid-desc.component';
+import { AcPolylineVolumeDescComponent } from './components/ac-polyline-volume-desc/ac-polyline-volume-desc.component';
+import { AcWallDescComponent } from './components/ac-wall-desc/ac-wall-desc.component';
+import { AcRectangleDescComponent } from './components/ac-rectangle-desc/ac-rectangle-desc.component';
 
 @NgModule({
 	imports: [
 		CommonModule,
 		Angular2ParseModule,
-		UtilsModule
+		UtilsModule,
 	],
 	declarations: [
 		AcMapComponent,
@@ -71,6 +79,14 @@ import { AcModelDescComponent } from './components/ac-model-desc/ac-model-desc.c
     AcPolygonComponent,
     AcDefaultPlonterComponent,
 		AcModelDescComponent,
+    AcTileset3dComponent,
+		AcBoxDescComponent,
+		AcCylinderDescComponent,
+		AcCorridorDescComponent,
+		AcEllipsoidDescComponent,
+		AcPolylineVolumeDescComponent,
+		AcWallDescComponent,
+		AcRectangleDescComponent,
     
     AcStaticEllipseDescComponent,
   	AcDynamicEllipseDescComponent,
@@ -104,8 +120,16 @@ import { AcModelDescComponent } from './components/ac-model-desc/ac-model-desc.c
     AcPolygonComponent,
     AcDefaultPlonterComponent,
 		AcModelDescComponent,
-    
-    AcStaticEllipseDescComponent,
+    AcTileset3dComponent,
+		AcBoxDescComponent,
+		AcCylinderDescComponent,
+		AcCorridorDescComponent,
+		AcEllipsoidDescComponent,
+		AcPolylineVolumeDescComponent,
+		AcWallDescComponent,
+		AcRectangleDescComponent,
+		
+		AcStaticEllipseDescComponent,
     AcDynamicEllipseDescComponent,
     AcDynamicPolylineDescComponent,
     AcStaticPolylineDescComponent,