+++ src/models/layer-options.ts
@@ -4,4 +4,12 @@ export interface LayerOptions {
   ellipse?: EntitiesDrawerOptions;
   circle?: EntitiesDrawerOptions;
   polygon?: EntitiesDrawerOptions;
+	model?: EntitiesDrawerOptions;
+	box?: EntitiesDrawerOptions;
+	corridor?: EntitiesDrawerOptions;
+	cylinder?: EntitiesDrawerOptions;
+	ellipsoid?: EntitiesDrawerOptions;
+	polylineVolume?: EntitiesDrawerOptions;
+	rectangle?: EntitiesDrawerOptions;
+	wall?: EntitiesDrawerOptions;
 }