+++ CHANGELOG.md
@@ -1,12 +1,22 @@
-## 0.0.12 (30-7-17)
-### Breaking changes
-* AcEntity id filed changed to string from number
-* AcNotification id filed changed to string from number
-
-## 0.0.13
+## 0.0.16
 ### Features
-* ac-map-layer-provider - Support all cesium imagery providers
-* Expose cesium viewer through ViewesManagerService and `ac-map`
+* Expose `onDraw` and `onRemove` for `BasicDesc` component - now it's possible to know when an Cesium object is drawn.
+* Expose `getCesiumObjectsMap` from `BasicDesc` component.
+* Add drag and drop event to `MapEventManagerService`
+* support:
+  * `ac-box-dec`
+  * `ac-corridor-dec`
+  * `ac-cylinder-dec`
+  * `ac-ellipsoid-dec`
+  * `ac-polyline-volume-dec`
+  * `ac-wall-dec`
+  * `ac-rectangle-dec`
+
+## 0.0.15
+### Features
+* Expose `MapEventManagerService` from `<ac-map #map/>` component reference: `acMapComponent.getMapEventManager()`
+* New `KeyboardControlService` for controlling the camera using the keyboard
+* add `<ac-3d-tile-layer/>` for 3d tiles
 
 ## 0.0.14
 ### Breaking changes 
@@ -21,3 +31,13 @@
 
 ### Features 
 * support models `ac-model-desc`
+
+## 0.0.13
+### Features
+* ac-map-layer-provider - Support all cesium imagery providers
+* Expose cesium viewer through ViewesManagerService and `ac-map`
+
+## 0.0.12 (30-7-17)
+### Breaking changes
+* AcEntity id filed changed to string from number
+* AcNotification id filed changed to string from number